ajout key event + CBopcode handler
