+++ projectSrc/GBmu.pro
@@ -24,7 +24,8 @@ SOURCES  += src/main.cpp \
 			src/gb/core/Gpu.cpp \
 			src/gb/core/Timer.cpp \
 			src/gb/Gameboy.cpp \
-			src/gb/core/CpuOpcode.cpp
+			src/gb/core/CpuOpcode.cpp \
+			src/gb/core/CpuCbOpcode.cpp
 
 
 HEADERS  += include/gb/gui/OpenGLWindow.hpp \
