+++ projectSrc/include/gb/core/Cpu.hpp
@@ -63,6 +63,7 @@ class Cpu_z80
 		uint8_t							_getLengthDataOpcode(void);
 		void							_setDataOpcode(void);
 		void							_setOpcodeMap(void);
+		void							_setCbOpcodeMap(void);
 
 	/*
 	** ################################################################
@@ -75,6 +76,11 @@ class Cpu_z80
 		void							interrupt(void);
 		void							init(void);
 
+	/*
+	** ################################################################
+	** METHOD Opcode
+	** ################################################################
+	*/
 
 		void							XOR(uint8_t val);
 		void							AND(uint8_t val);
@@ -335,7 +341,15 @@ class Cpu_z80
 		void							EI();			//0xfb     Not done yet
 		void							CP_n();			//0xfe 
 		void							RST_38H();		//0xff  		
-				
+
+	/*
+	** ################################################################
+	** METHOD CB opcode
+	** ################################################################
+	*/
+		void							rotate_left(uint8_t *reg);
+		void							RLC_B();		//0x00
+	
 	private:
 		void							_resetPtrAddr(void);
 		void							_nextPtr(void);
